Factor out test crate for use in benchmarks
