+++ x/contracts/wasmlanche/benches/call_contract.rs
@@ -1,0 +1,1 @@
+
